requirements.txt
aiohttp==3.8.1
bleak==0.14.0
crc16==0.1.1
PyNaCl==1.3.0
PyYAML==6.0
